Fix the permissions of /usr/bin/sudo in the target image.
For some reason the COPY instruction does not preserve the setuid or setgid
bits when it copies files into the target. Without this, the unprivileged user
cannot sudo properly, and thus can't install new tools in their container.

In containers/buildah/issues/574 it is suggested this may be a bug or at least
an undocumented side-effect of permission blitting in Go, but for us it is easy
enough to just fix up the one file we care about.