Update volume setup for LLVM as an external dependency.
kythe/kythe/pull/3422 replaced the script-driven LLVM build with a normal
external dependency that builds with Bazel. This means we can no longer
practically keep LLVM in a separate volume, as it is now folded together with
the other Bazel outputs.

Instead, rework the LLVM volume as a more general Bazel cache volume, and set
up ~/.bazelrc to include a --disk_cache flag pointing to it. This has the same
practical benefit of speeding up re-builds of LLVM, and in fact helps with any
build because it's shared.