Fix Bazel installation to allow pinning a specific version.
The Bazel apt repository only advertises the latest release, so when Bazel
ships breaking changes the image may pull in a version that is not compatible
with Kythe's macros and custom rules. To make that easier to address, pin a
specific version, and fetch the binary release instead.

This also winds up simplifiying the dependency setup, since the published .deb
file already knows what dependencies to grab.