Copy the Go installation from the official images.
Rather than fetching and installing a tarball, pull the golang image for the
desired version series and copy the installation from it. This turns out to be
quite a bit faster even accounting for the fetch, and the installation is all
self-contained so we don't pick up a lot of ancillary detritus.